icons: introduce v3
