+++ packages/design-system/components/button/__stories__/Button.stories.mdx
@@ -26,7 +26,7 @@ Allows users to take actions on a website. It can be a regular button, a link or
     name="Overview - button icon end"
     args={{
       text: "Button",
-      iconEnd: "arrowDown",
+      iconEnd: "arrowRight",
       variant: "fill",
       size: "medium",
     }}
