+++ packages/design-system/components/icon/index.js
@@ -1,4 +1,3 @@
 import Icon from "./Icon"
 
-export { icons } from "./Icon"
 export default Icon
