+++ packages/design-system/components/list/__stories__/List.stories.mdx
@@ -26,10 +26,10 @@ Displays list items. It renders a `<ul>` element by default.
 
 export const Template = (args) => (
   <List {...args}>
-    <ListItem iconName="done">Lista element 1 </ListItem>
-    <ListItem iconName="done" variant="gray">
+    <ListItem iconName="check">Lista element 1 </ListItem>
+    <ListItem iconName="check" variant="gray">
       Lista element 2
     </ListItem>
-    <ListItem iconName="done">Lista element 3 </ListItem>
+    <ListItem iconName="check">Lista element 3 </ListItem>
   </List>
 )
