+++ packages/design-system/components/listItem/__stories__/ListItem.stories.mdx
@@ -16,7 +16,7 @@ Displays a list item element, as a `li` tag.
     args={{
       children: "umożliwiamy adopcję za granicę",
       variant: "primary",
-      iconName: "done",
+      iconName: "check",
     }}
   >
     {Template.bind()}
