+++ packages/design-system/patterns/textBanner/__stories__/TextBanner.stories.mdx
@@ -25,7 +25,7 @@ Displays a group of text elements in a suitable layout.
           text="Button"
           variant="text"
           color="primary"
-          iconStart="arrowDown"
+          iconEnd="arrowRight"
         />
       ),
       href: "https://google.com",
