+++ packages/design-system/tokens/icons/__stories__/Icons.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, IconGallery, IconItem } from "@storybook/addon-docs"
 
-import { icons } from "design-system/components/icon"
+import { icons } from "../"
 
 <Meta title="Tokens/Icons" />
