+++ packages/design-system/tokens/icons/index.js
@@ -1,0 +1,69 @@
+import ArrowLeft from "./ArrowLeft.svg"
+import ArrowRight from "./ArrowRight.svg"
+import Cat from "./Cat.svg"
+import Charity from "./Charity.svg"
+import Check from "./Check.svg"
+import ChevronLeft from "./ChevronLeft.svg"
+import ChevronRight from "./ChevronRight.svg"
+import ChevronUp from "./ChevronUp.svg"
+import ChevronDown from "./ChevronDown.svg"
+import Dog from "./Dog.svg"
+import Download from "./Download.svg"
+import Facebook from "./Facebook.svg"
+import Heart from "./Heart.svg"
+import Instagram from "./Instagram.svg"
+import Location from "./Location.svg"
+import Mail from "./Mail.svg"
+import Phone from "./Phone.svg"
+import Close from "./Close.svg"
+import Paperclip from "./Paperclip.svg"
+import Bowl from "./Bowl.svg"
+import Menu from "./Menu.svg"
+
+export const icons = {
+  arrowLeft: ArrowLeft,
+  arrowRight: ArrowRight,
+  check: Check,
+  chevronLeft: ChevronLeft,
+  chevronRight: ChevronRight,
+  chevronUp: ChevronUp,
+  chevronDown: ChevronDown,
+  close: Close,
+  download: Download,
+  facebook: Facebook,
+  instagram: Instagram,
+  location: Location,
+  phone: Phone,
+  mail: Mail,
+  charity: Charity,
+  heart: Heart,
+  cat: Cat,
+  dog: Dog,
+  paperClip: Paperclip,
+  bowl: Bowl,
+  menu: Menu,
+}
+
+export {
+  ArrowLeft,
+  ArrowRight,
+  Bowl,
+  Cat,
+  Check,
+  Charity,
+  ChevronLeft,
+  ChevronRight,
+  ChevronUp,
+  ChevronDown,
+  Dog,
+  Download,
+  Facebook,
+  Menu,
+  Heart,
+  Instagram,
+  Location,
+  Mail,
+  Phone,
+  Close,
+  Paperclip,
+}
